requirements.txt
certifi==2021.5.30
chardet==4.0.0
Django==2.2.18
django-environ==0.4.5
django-ranged-response==0.2.0
django-simple-captcha==0.5.14
idna==2.10
Pillow==8.2.0
psycopg2==2.8.6
pytz==2021.1
requests==2.25.1
six==1.16.0
sqlparse==0.4.1
urllib3==1.26.5
